Use webcrypto based connect-js library.
This adopts usage of the WebCrypto Api based connect-js library
as tracked on
https://github.com/anvilresearch/connect-js/issues/7#issuecomment-175079961
With this most API methods are now returning promises.

Summary of other  changes:

* Use npm and browserify to consume connect-js
* Adopt latest anvil cli

The changes were made over a period of time. Here are some of the more
notable original commits with comments:

COMMIT: callback_popup to check whether opener is available
COMMIT: index.html shows userInfo.name may not be available
COMMIT: add APP_AUTH_CALLBACK variable
COMMIT: establish scope.session in SigninCtrl and MainCtrl
COMMIT: scope to requireScope profile instead of realm
COMMIT: adopt new cli
COMMIT: Handle disconnected popup in passwordless login.

The passwordless login method sends the user a link in an email. When
the user follows this link a new browser window opens.

However the original popup window will not redirect and remains open
with a page allowing to resend the link.

The connect-js library was changed to handle this case so that it would
expect the page which is opened when the email link is followed
to handle the callback so that the session is populated appropriately
in localStorage.

connect-js now listens for that and then closes the popup window.
It also expect the email page to close itself.

This may not work for all browsers.  I did see this work in Chrome and Safari
but fail in Firefox.
Note that the popup Callback should behave in the same way it it is
redirected by the server, which happens when other authentication
methods are used.

This also adds the newer dependencies and improves logging similar to
how it was done in connect-js.

COMMIT: refresh angular app on passwordless login
COMMIT: adding jspm bundle-sfx for demo

COMMIT: Manual angular bootstrap

This was to avoid issue https://github.com/systemjs/systemjs/issues/1032

COMMIT: Switch to using npm and browserify

This now requires anvil-connect-js ^0.2

bower is still used for shims and bootstrap.

Some bugs were addressed found during manual testing.
Also added more logging using bows.

COMMIT: Added section Get connect-js client libraries

I am uncertain about the remainder of the doc, but it looked OK to me at
first glance although I have not done a client registration for quite some
time.